# packages/vantage/src/vantage/service/openapi/v1.yaml
...
openapi: 3.1.0
info:
  title: Vantage API
  version: "1"
  description: read-tagged paths are guaranteed to leave stored data unchanged; write-tagged paths change stored data. Every rejection -- each 4xx and 5xx listed here, and the 404 or 405 for a path nothing serves or a method a path does not take -- has the Rejection body.
servers:
  - url: /api/v1
...
paths:
  /runs:
    post:
      operationId: create_run
      tags: [write]
      summary: Report a session's run and its results.
      description: Declared metadata never rejects a report. The server bounds it the way the plugin does -- it drops a file repeating an earlier file's path, over-long and repeated keys and every key past the per-run entry bound, and records a document past the per-file or per-report size as too_large or over_budget without parsing it.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SessionReport'}}}}
      responses:
        "201": {description: "Stored, a new run.", content: {application/json: {schema: {$ref: '#/components/schemas/Acknowledgement'}}}}
        "200": {description: "Retried; already stored (idempotent replay).", content: {application/json: {schema: {$ref: '#/components/schemas/Acknowledgement'}}}}
        "400": {description: "The body is not UTF-8 JSON (invalid_json), or the client disconnected before sending all of it (incomplete_body).", content: &rejection {application/json: {schema: {$ref: '#/components/schemas/Rejection'}}}}
        "413": {description: "The body exceeds 1 MiB (payload_too_large).", content: *rejection}
        "415": {description: "Content-Type is not application/json (unsupported_media_type).", content: *rejection}
        "422": {description: "The body does not match SessionReport (invalid_report); fields names each failing field.", content: *rejection}
    get:
      operationId: list_runs
      tags: [read]
      summary: Page through recorded runs, newest first.
      description: metadata_key and metadata_value together filter to runs holding that declared key=value pair -- both or neither, one without the other is 422. A run recorded before the key was ever declared has no value for it and is excluded; metadata_horizon in the response reports how many runs predate it (null when no metadata filter was given).
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - {name: metadata_key, in: query, required: false, schema: {type: string}}
        - {name: metadata_value, in: query, required: false, schema: {type: string}}
      responses:
        "200": {description: A page of runs., content: {application/json: {schema: {$ref: '#/components/schemas/RunListResponse'}}}}
        "422": {description: "limit below 1 or offset out of range -- not a page; or metadata_key/metadata_value supplied without its pair.", content: *rejection}
  /runs/{run_id}:
    get:
      operationId: get_run_detail
      tags: [read]
      summary: The full record for one run, with its untruncated commit subject.
      parameters:
        - $ref: '#/components/parameters/run_id'
      responses:
        "200": {description: The run., content: {application/json: {schema: {$ref: '#/components/schemas/RunDetailResponse'}}}}
        "404": {description: No run with that id has been recorded., content: *rejection}
        "422": {description: run_id is not 32 lowercase hex characters., content: *rejection}
  /runs/{run_id}/heartbeat:
    post:
      operationId: heartbeat
      tags: [write]
      summary: Advance a run's last-contact timestamp.
      parameters:
        - $ref: '#/components/parameters/run_id'
      responses:
        "200": {description: Acknowledged., content: {application/json: {schema: {$ref: '#/components/schemas/HeartbeatAcknowledgement'}}}}
        "404": {description: No run with that id has been recorded., content: *rejection}
        "422": {description: run_id is not 32 lowercase hex characters., content: *rejection}
...
components:
  parameters:
    run_id: {name: run_id, in: path, required: true, schema: {type: string, pattern: '^[0-9a-f]{32}$'}}
...
    limit: {name: limit, in: query, required: false, description: "Page size. Values above 200 are clamped to 200; the response reports whether more items exist. A value below 1 is rejected with 422 -- it is not a page size.", schema: {type: integer, minimum: 1}}
    offset: {name: offset, in: query, required: false, schema: {type: integer, format: int64, minimum: 0}}
  schemas:
...
    SessionReport:
      type: object
      required: [run]
      properties: {run: {$ref: '#/components/schemas/RunReport'}, results: {type: [array, "null"], items: {$ref: '#/components/schemas/ResultReport'}}, vcs: {oneOf: [{$ref: '#/components/schemas/VcsReport'}, {type: "null"}]}, metadata: {oneOf: [{$ref: '#/components/schemas/MetadataReport'}, {type: "null"}]}}
    RunReport:
      type: object
      required: [id, started_at, finished_at, exit_status, interrupted, interrupt_reason]
      additionalProperties: false
      properties: {id: {type: string, pattern: '^[0-9a-f]{32}$'}, started_at: {type: string, format: date-time}, finished_at: {type: [string, "null"], format: date-time}, exit_status: {type: [integer, "null"], format: int64}, interrupted: {type: boolean}, interrupt_reason: {type: [string, "null"]}}
    # An unknown key on one result is tolerated and named back in
    # `Acknowledgement.ignored`; the failure and captured-output fields are
    # optional, because an older plugin sends none of them.
    ResultReport:
      type: object
      required: [node_id, file_path, class_name, function_name, param_id, outcome, duration, started_at, finished_at, setup_outcome, call_outcome, teardown_outcome, setup_duration, call_duration, teardown_duration, worker_id]
      additionalProperties: true
      properties: {node_id: {type: string}, file_path: {type: string}, class_name: {type: [string, "null"]}, function_name: {type: string}, param_id: {type: [string, "null"]}, outcome: {type: string, enum: [passed, failed, error, skipped, xfailed, xpassed]}, duration: {type: [number, "null"]}, started_at: {type: [string, "null"], format: date-time}, finished_at: {type: [string, "null"], format: date-time}, setup_outcome: {type: [string, "null"], enum: [passed, failed, error, skipped, xfailed, xpassed, null]}, call_outcome: {type: [string, "null"], enum: [passed, failed, error, skipped, xfailed, xpassed, null]}, teardown_outcome: {type: [string, "null"], enum: [passed, failed, error, skipped, xfailed, xpassed, null]}, setup_duration: {type: [number, "null"]}, call_duration: {type: [number, "null"]}, teardown_duration: {type: [number, "null"]}, worker_id: {type: [string, "null"]}, failure_type: {type: [string, "null"], default: null}, failure_message: {type: [string, "null"], default: null}, failure_message_truncated: {type: boolean, default: false}, failure_path: {type: [string, "null"], default: null}, failure_lineno: {type: [integer, "null"], format: int64, default: null}, failure_repr: {type: [string, "null"], default: null}, failure_repr_truncated: {type: boolean, default: false}, traceback: {type: [string, "null"], default: null}, traceback_truncated: {type: boolean, default: false}, skip_reason: {type: [string, "null"], default: null}, skip_reason_truncated: {type: boolean, default: false}, xfail_reason: {type: [string, "null"], default: null}, xfail_reason_truncated: {type: boolean, default: false}, captured_stdout: {type: [string, "null"], default: null}, captured_stdout_truncated: {type: boolean, default: false}, captured_stderr: {type: [string, "null"], default: null}, captured_stderr_truncated: {type: boolean, default: false}}
    VcsReport:
      type: object
      required: [commit, branch, commit_subject, dirty, root]
      additionalProperties: false
      properties: {commit: {type: [string, "null"], maxLength: 64}, branch: {type: [string, "null"]}, commit_subject: {type: [string, "null"]}, dirty: {type: [boolean, "null"]}, root: {type: [string, "null"]}}
    MetadataReport:
      type: object
      required: [declaration, files]
      additionalProperties: false
      properties: {declaration: {type: string}, files: {type: array, items: {$ref: '#/components/schemas/MetadataFileReport'}}}
    # No constraint on any field: a declared value is text this project does
    # not control, so the server drops what it cannot store rather than
    # rejecting the report. `content` is null unless `status` is `captured`.
    MetadataFileReport:
      type: object
      required: [path, format, status, keys, content]
      additionalProperties: false
      properties: {path: {type: string}, format: {type: string}, status: {type: string}, keys: {type: array, items: {type: string}}, content: {type: [string, "null"]}}
    # The one body every rejection has. `detail` is a fixed sentence and
    # `fields` holds dotted field paths; neither ever repeats a submitted
    # value.
    Rejection:
      type: object
      required: [error, detail, fields]
      properties: {error: {type: string}, detail: {type: string}, fields: {type: array, items: {type: string}}}
    Acknowledgement:
      type: object
      required: [run_id, status, ignored]
      properties: {run_id: {type: string}, status: {type: string}, ignored: {type: array, items: {type: string}}}
    HeartbeatAcknowledgement:
      type: object
      required: [run_id, status]
      properties: {run_id: {type: string}, status: {type: string}}
    # No `root`. `VcsContext` carries one; this model has no field to
    # populate, which is what keeps a reporter's absolute path off the wire
    # on the detail path where no type stops it.
    RunVcs:
      type: object
      required: [commit, branch, commit_subject, commit_subject_truncated, dirty]
      properties: {commit: {type: [string, "null"]}, branch: {type: [string, "null"]}, commit_subject: {type: [string, "null"]}, commit_subject_truncated: {type: boolean}, dirty: {type: [boolean, "null"]}}
    RunListItem:
      type: object
      required: [id, started_at, finished_at, exit_status, interrupted, presentation, vcs]
      properties: {id: {type: string, pattern: '^[0-9a-f]{32}$'}, started_at: {type: string, format: date-time}, finished_at: {type: [string, "null"], format: date-time}, exit_status: {type: [integer, "null"]}, interrupted: {type: boolean}, presentation: {type: string, enum: [running, finished, interrupted, abandoned]}, vcs: {oneOf: [{$ref: '#/components/schemas/RunVcs'}, {type: "null"}]}}
    # `predating` equals the total run count when `key` was never declared.
    # Populated only when a metadata filter was given.
    MetadataHorizon:
      type: object
      required: [key, predating]
      properties: {key: {type: string}, predating: {type: integer}}
    RunListResponse:
      type: object
      required: [items, has_more, metadata_horizon]
      properties: {items: {type: array, items: {$ref: '#/components/schemas/RunListItem'}}, has_more: {type: boolean}, metadata_horizon: {oneOf: [{$ref: '#/components/schemas/MetadataHorizon'}, {type: "null"}]}}
    # Carries `interrupt_reason` and an untruncated `commit_subject`, which
    # the lean list entry omits -- the detail path keeps the full record
    # reachable.
    RunDetailResponse:
      type: object
      required: [id, started_at, finished_at, exit_status, interrupted, interrupt_reason, presentation, vcs]
      properties: {id: {type: string, pattern: '^[0-9a-f]{32}$'}, started_at: {type: string, format: date-time}, finished_at: {type: [string, "null"], format: date-time}, exit_status: {type: [integer, "null"]}, interrupted: {type: boolean}, interrupt_reason: {type: [string, "null"]}, presentation: {type: string, enum: [running, finished, interrupted, abandoned]}, vcs: {oneOf: [{$ref: '#/components/schemas/RunVcs'}, {type: "null"}]}}
    # No traceback, failure_repr or captured output at any depth -- a
    # results list never carries them; the single-result endpoint does.
...